Refactor functions to own modules
